requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.4.2
pyodbc==4.0.34
SpeechRecognition==3.8.1
streamlit==1.12.0
